Fix the regcache in case of several pending requests at the same time: * we can't have multiple rndv pending using the same region since the   protocol only returns a 8bit id to find out which request to complete.   add a "reserver" field to the region to prevent the regcache from   using it again before the notify arrives. * each region may be used by multiple pull at the same time as long   as we can translate a pull done event into a large recv to complete.   pass the request in the pull request/event instead of the region id   to do so. * drop the user field from the request now, since it is either multiple   pull users, or a single rndv/notify reserver.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1110 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356

+++ driver/linux/omx_pull.c
@@ -105,7 +105,7 @@ struct omx_pull_handle {
 	/* global pull fields */
 	struct omx_endpoint * endpoint;
 	struct omx_user_region * region;
-	uint32_t lib_cookie;
+	uint64_t lib_cookie;
 	uint32_t total_length;
 	uint32_t puller_rdma_offset;
 	uint32_t pulled_rdma_offset;
